app/requirements.txt
flet==0.22.0
Pillow==10.3.0
pystray==0.19.5
watchdog==4.0.0
